refactor(jina): 더보기 버튼을 공통 컴포넌트로 적용 #45
